Fix CodeQL SM02324 bad overflow checks in HII code
Cast UINT16 Offset+Width additions to UINT32 to prevent overflow
before comparison in block merge and overlap detection logic.

+++ MdeModulePkg/Library/UefiHiiLib/HiiLib.c
@@ -2120,9 +2120,9 @@ GetBlockDataInfo (
   while ((Link != &BlockArray->Entry) && (Link->ForwardLink != &BlockArray->Entry)) {
     BlockData    = BASE_CR (Link, IFR_BLOCK_DATA, Entry);
     NewBlockData = BASE_CR (Link->ForwardLink, IFR_BLOCK_DATA, Entry);
-    if ((NewBlockData->Offset >= BlockData->Offset) && (NewBlockData->Offset <= (BlockData->Offset + BlockData->Width))) {
-      if ((NewBlockData->Offset + NewBlockData->Width) > (BlockData->Offset + BlockData->Width)) {
-        BlockData->Width = (UINT16)(NewBlockData->Offset + NewBlockData->Width - BlockData->Offset);
+    if ((NewBlockData->Offset >= BlockData->Offset) && (NewBlockData->Offset <= (UINT32)BlockData->Offset + BlockData->Width)) {
+      if ((UINT32)NewBlockData->Offset + NewBlockData->Width > (UINT32)BlockData->Offset + BlockData->Width) {
+        BlockData->Width = (UINT16)((UINT32)NewBlockData->Offset + NewBlockData->Width - BlockData->Offset);
       }
 
       RemoveEntryList (Link->ForwardLink);
